# Bugfix for BufferedTokenizer to completely consume lines in case of lines bigger then sizeLimit