@andsel 2024-10-16
### Benchmark

Result with performance metering of the `extract` method using the benchmark https://github.com/elastic/logstash/pull/16564/files

| test name                               | baseline (ops/ns) | bugfix (ops/ns) | delta |
|-----------------------------------------|------------------:|-----------------|-------|
| onlyOneTokenPerFragment                 | 13702             | 7982            | -41%  |
| multipleTokenPerFragment                | 4276              | 3553            | -16%  |
| multipleTokensCrossingMultipleFragments | 2180              | 1420            | -34%  |


<details>
  <summary>Raw baseline results</summary>
  <p>
  
```
andrea:logstash_andsel (main) % ./gradlew jmh -Pinclude="org.logstash.benchmark.BufferedTokenizerExtBenchmark.*"
To honour the JVM settings for this build a single-use Daemon process will be forked. For more on this, please refer to https://docs.gradle.org/8.7/userguide/gradle_daemon.html#sec:disabling_the_daemon in the Gradle documentation.
Daemon will be stopped at the end of the build

> Task :downloadJRuby UP-TO-DATE
Download https://repo1.maven.org/maven2/org/jruby/jruby-dist/9.4.8.0/jruby-dist-9.4.8.0-bin.tar.gz

> Task :logstash-core:compileJava
Note: Processing Log4j annotations
Note: Annotations processed
Note: Processing Log4j annotations
Note: No elements to process

> Task :logstash-core-benchmarks:jmh
# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokenPerFragment

# Run progress: 0.00% complete, ETA 00:00:03
# Fork: 1 of 1
# Warmup Iteration   1: 56.913 ops/ns
# Warmup Iteration   2: 2636.928 ops/ns
# Warmup Iteration   3: 3882.360 ops/ns
Iteration   1: 4327.280 ops/ns
Iteration   2: 4267.082 ops/ns
Iteration   3: 4371.703 ops/ns
Iteration   4: 4252.072 ops/ns
Iteration   5: 4244.560 ops/ns
Iteration   6: 4245.370 ops/ns[10s]
Iteration   7: 4222.432 ops/ns[11s]
Iteration   8: 4197.453 ops/ns[11s]
Iteration   9: 4315.799 ops/ns[11s]
Iteration  10: 4316.202 ops/ns[11s]


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokenPerFragment":
  4275.995 ±(99.9%) 82.202 ops/ns [Average]
  (min, avg, max) = (4197.453, 4275.995, 4371.703), stdev = 54.372
  CI (99.9%): [4193.793, 4358.198] (assumes normal distribution)


# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments

# Run progress: 33.33% complete, ETA 00:00:04
# Fork: 1 of 1
# Warmup Iteration   1: 49.464 ops/ns
# Warmup Iteration   2: 1222.981 ops/ns
# Warmup Iteration   3: 2031.913 ops/ns
Iteration   1: 2184.758 ops/ns[12s]
Iteration   2: 2231.615 ops/ns[12s]
Iteration   3: 2211.407 ops/ns[12s]
Iteration   4: 2264.482 ops/ns[12s]
Iteration   5: 2172.669 ops/ns[13s]
Iteration   6: 2110.849 ops/ns[13s]
Iteration   7: 2161.588 ops/ns[13s]
Iteration   8: 2151.354 ops/ns[13s]
Iteration   9: 2088.678 ops/ns[13s]
Iteration  10: 2221.383 ops/ns[13s]


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments":
  2179.878 ±(99.9%) 82.590 ops/ns [Average]
  (min, avg, max) = (2088.678, 2179.878, 2264.482), stdev = 54.628
  CI (99.9%): [2097.289, 2262.468] (assumes normal distribution)


# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment

# Run progress: 66.67% complete, ETA 00:00:02
# Fork: 1 of 1
# Warmup Iteration   1: 82.220 ops/ns
# Warmup Iteration   2: 3971.738 ops/ns
# Warmup Iteration   3: 9871.581 ops/ns
Iteration   1: 13820.481 ops/ns14s]
Iteration   2: 14055.263 ops/ns15s]
Iteration   3: 14028.545 ops/ns15s]
Iteration   4: 12603.806 ops/ns15s]
Iteration   5: 13157.192 ops/ns15s]
Iteration   6: 13159.603 ops/ns15s]
Iteration   7: 13454.914 ops/ns15s]
Iteration   8: 14170.723 ops/ns15s]
Iteration   9: 14245.842 ops/ns
Iteration  10: 14325.432 ops/ns


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment":
  13702.180 ±(99.9%) 873.989 ops/ns [Average]
  (min, avg, max) = (12603.806, 13702.180, 14325.432), stdev = 578.090
  CI (99.9%): [12828.191, 14576.169] (assumes normal distribution)


# Run complete. Total time: 00:00:06

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

NOTE: Current JVM experimentally supports Compiler Blackholes, and they are in use. Please exercise
extra caution when trusting the results, look into the generated code to check the benchmark still
works, and factor in a small probability of new VM bugs. Additionally, while comparisons between
different JVMs are already problematic, the performance difference caused by different Blackhole
modes can be very significant. Please make sure you use the consistent Blackhole mode for comparisons.

Benchmark                                                               Mode  Cnt      Score     Error   Units
BufferedTokenizerExtBenchmark.multipleTokenPerFragment                 thrpt   10   4275.995 ±  82.202  ops/ns
BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments  thrpt   10   2179.878 ±  82.590  ops/ns
BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment                  thrpt   10  13702.180 ± 873.989  ops/ns
```

  </p>
</details>

<details>
  <summary>Bugfix results</summary>
  <p>

```
andrea:logstash_andsel (fix/buffered_tokenizer_clean_state_in_case_of_line_too_big) % ./gradlew jmh -Pinclude="org.logstash.benchmark.BufferedTokenizerExtBenchmark.*"
To honour the JVM settings for this build a single-use Daemon process will be forked. For more on this, please refer to https://docs.gradle.org/8.7/userguide/gradle_daemon.html#sec:disabling_the_daemon in the Gradle documentation.
Daemon will be stopped at the end of the build

> Task :downloadJRuby UP-TO-DATE
Download https://repo1.maven.org/maven2/org/jruby/jruby-dist/9.4.8.0/jruby-dist-9.4.8.0-bin.tar.gz

> Task :logstash-core-benchmarks:jmh
# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokenPerFragment

# Run progress: 0.00% complete, ETA 00:00:03
# Fork: 1 of 1
# Warmup Iteration   1: 38.059 ops/ns
# Warmup Iteration   2: 1590.637 ops/ns
# Warmup Iteration   3: 2992.422 ops/ns
Iteration   1: 3715.554 ops/ns
Iteration   2: 3963.687 ops/ns
Iteration   3: 3780.469 ops/ns
Iteration   4: 3339.686 ops/ns
Iteration   5: 3329.894 ops/ns
Iteration   6: 3281.113 ops/ns
Iteration   7: 3476.390 ops/ns[9s]
Iteration   8: 3753.612 ops/ns[9s]
Iteration   9: 3488.085 ops/ns[9s]
Iteration  10: 3407.034 ops/ns[9s]


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokenPerFragment":
  3553.552 ±(99.9%) 351.924 ops/ns [Average]
  (min, avg, max) = (3281.113, 3553.552, 3963.687), stdev = 232.776
  CI (99.9%): [3201.628, 3905.477] (assumes normal distribution)


# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments

# Run progress: 33.33% complete, ETA 00:00:04
# Fork: 1 of 1
# Warmup Iteration   1: 25.704 ops/ns
# Warmup Iteration   2: 789.568 ops/ns
# Warmup Iteration   3: 1413.859 ops/ns
Iteration   1: 1596.075 ops/ns[10s]
Iteration   2: 1490.031 ops/ns[10s]
Iteration   3: 1572.851 ops/ns[10s]
Iteration   4: 1410.213 ops/ns[10s]
Iteration   5: 1418.818 ops/ns[11s]
Iteration   6: 1406.862 ops/ns[11s]
Iteration   7: 1091.753 ops/ns[11s]
Iteration   8: 1529.710 ops/ns[11s]
Iteration   9: 1388.452 ops/ns[11s]
Iteration  10: 1298.760 ops/ns[11s]


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments":
  1420.352 ±(99.9%) 222.548 ops/ns [Average]
  (min, avg, max) = (1091.753, 1420.352, 1596.075), stdev = 147.202
  CI (99.9%): [1197.804, 1642.901] (assumes normal distribution)


# JMH version: 1.37
# VM version: JDK 21.0.4, OpenJDK 64-Bit Server VM, 21.0.4+7-LTS
# VM invoker: /Users/andrea/.sdkman/candidates/java/21.0.4-tem/bin/java
# VM options: -Dfile.encoding=UTF-8 -Duser.country=IT -Duser.language=en -Duser.variant -Djava.io.tmpdir=/Users/andrea/workspace/logstash_andsel/logstash-core/benchmarks/build -XX:+HeapDumpOnOutOfMemoryError -Xms2g -Xmx2g
# Blackhole mode: compiler (auto-detected, use -Djmh.blackhole.autoDetect=false to disable)
# Warmup: 3 iterations, 100 ms each
# Measurement: 10 iterations, 100 ms each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: org.logstash.benchmark.BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment

# Run progress: 66.67% complete, ETA 00:00:02
# Fork: 1 of 1
# Warmup Iteration   1: 45.601 ops/ns
# Warmup Iteration   2: 2655.867 ops/ns
# Warmup Iteration   3: 5615.446 ops/ns
Iteration   1: 8124.816 ops/ns[12s]
Iteration   2: 8347.746 ops/ns[13s]
Iteration   3: 8412.099 ops/ns[13s]
Iteration   4: 6964.810 ops/ns[13s]
Iteration   5: 7227.234 ops/ns[13s]
Iteration   6: 7313.477 ops/ns[13s]
Iteration   7: 8263.972 ops/ns[13s]
Iteration   8: 8622.241 ops/ns[13s]
Iteration   9: 8165.904 ops/ns[13s]
Iteration  10: 8381.400 ops/ns[13s]


Result "org.logstash.benchmark.BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment":
  7982.370 ±(99.9%) 883.453 ops/ns [Average]
  (min, avg, max) = (6964.810, 7982.370, 8622.241), stdev = 584.349
  CI (99.9%): [7098.917, 8865.823] (assumes normal distribution)


# Run complete. Total time: 00:00:06

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

NOTE: Current JVM experimentally supports Compiler Blackholes, and they are in use. Please exercise
extra caution when trusting the results, look into the generated code to check the benchmark still
works, and factor in a small probability of new VM bugs. Additionally, while comparisons between
different JVMs are already problematic, the performance difference caused by different Blackhole
modes can be very significant. Please make sure you use the consistent Blackhole mode for comparisons.

Benchmark                                                               Mode  Cnt     Score     Error   Units
BufferedTokenizerExtBenchmark.multipleTokenPerFragment                 thrpt   10  3553.552 ± 351.924  ops/ns
BufferedTokenizerExtBenchmark.multipleTokensCrossingMultipleFragments  thrpt   10  1420.352 ± 222.548  ops/ns
BufferedTokenizerExtBenchmark.onlyOneTokenPerFragment                  thrpt   10  7982.370 ± 883.453  ops/ns
```

  </p>
</details>